+++ kpops/component_handlers/schema_handler/schema_handler.py
@@ -9,7 +9,7 @@ from schema_registry.client import AsyncSchemaRegistryClient
 from schema_registry.client.schema import AvroSchema
 
 from kpops.api.exception import ClassNotFoundError
-from kpops.cli.registry import find_class
+from kpops.api.registry import find_class
 from kpops.component_handlers.schema_handler.schema_provider import (
     Schema,
     SchemaProvider,
