+++ tests/cli/test_registry.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from kpops.cli.registry import ClassNotFoundError, Registry, _find_classes, find_class
+from kpops.api.registry import ClassNotFoundError, Registry, _find_classes, find_class
 from kpops.component_handlers.schema_handler.schema_provider import SchemaProvider
 from kpops.components.base_components.pipeline_component import PipelineComponent
 from tests.cli.resources.custom_module import CustomSchemaProvider
